Auto start from inside docker now works.
Changed auto_start.bash to use a function to start each process and to
create log files.
Added instructions to the docker/README.md file.
PX4 instance now starts as a daemon.
Stopped debug output from spawn_model.py.
Renamed connect.bash to attach.bash to reflect usage.
Removed drone/px4_reverse_engineer.md as they were my personal notes.
Fixed problem with drone package not being built.

+++ scripts/install/install_drone_packages.bash
@@ -9,4 +9,11 @@ set -e
 # Set up links to repos.
 mkdir -p ${COLCON_SRC_DIR}
 cd ${COLCON_SRC_DIR}
-ln -sf ${DRONE_GIT_DIR}
+
+# The drone repo is mounted on ~/code when inside a docker.
+if [ -f /.dockerenv ]
+then
+    ln -sf ~/code drone
+else
+    ln -sf ${DRONE_GIT_DIR}
+fi
